Create code
This code defines a rectangle object with length and width properties, then calculates its area and perimeter using the mathematical formulas for a rectangle. Finally, it prints the results to the console.

Note that this code assumes that the length and width of the rectangle are already defined in the rectangle object. If you wanted to prompt the user for these values, you could use the prompt() method to get input from the user and assign the values to the length and width properties of the rectangle object.